+++ NEWS.md
@@ -1,9 +1,18 @@
 # R Implementation of Advanced Optimizers for Torch
 
-### What's new in torchopt version 0.1.0
+### What's new in torchopt version 0.1.2
+
+* adahessian optimizer
+* nadam optimizer
+* radam optimizer 
+* qhadam optimizer
+* swats optimizer
+
+### What's new in torchopt version 0.1.1
 
 * adabelief optimizer
 * adabound optimizer
 * adamw optimizer
 * madgrad optimizer
 * yogi optimizer
+